Move sfu to pkg
