+++ cmd/signal/allrpc/server/server.go
@@ -7,7 +7,7 @@ import (
 	log "github.com/pion/ion-log"
 	grpcServer "github.com/pion/ion-sfu/cmd/signal/grpc/server"
 	jsonrpcServer "github.com/pion/ion-sfu/cmd/signal/json-rpc/server"
-	sfu "github.com/pion/ion-sfu/sfu"
+	"github.com/pion/ion-sfu/pkg/sfu"
 
 	// pprof
 	_ "net/http/pprof"
