+++ pkg/sfu/errors.go
