+++ pkg/sfu/router.go
@@ -200,14 +200,13 @@ func (r *router) loopDownTrackRTCP(track *DownTrack) {
 				}
 			case *rtcp.TransportLayerNack:
 				log.Tracef("sender got nack: %+v", p)
-				for _, pair := range p.Nacks {
-					r.wp.Submit(func() {
-						ps := r.buffer.GetBufferedPackets(track.receiver.SSRC(track.currentSpatialLayer), track.snOffset, track.tsOffset, track.nList.getNACKSeqNo(pair.PacketList()))
-						for _, pt := range ps {
+				r.wp.Submit(func() {
+					for _, pair := range p.Nacks {
+						for _, pt := range r.buffer.GetBufferedPackets(track.receiver.SSRC(track.currentSpatialLayer), track.snOffset, track.tsOffset, track.nList.getNACKSeqNo(pair.PacketList())) {
 							_ = track.WriteRTP(&pt)
 						}
-					})
-				}
+					}
+				})
 			}
 		}
 		if len(fwdPkts) > 0 {
